Moved the folder TimeEntries to the folder TimeRegistration. Moved User.cs from the Models folder to the Users folder. Moved Token.cs from the Models folder to the AccessTokens folder. Moved VacationOVerviewDto from the TimeEntries folder to the Absence folder. Changed the response in api/user/Profile fra User (from persistance) to UserResponseDto.
