+++ packages/api/AlvTime.Business/AccessTokens/IAccessTokenStorage.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using AlvTime.Business.Models;
+using AlvTime.Business.Users;
 
 namespace AlvTime.Business.AccessTokens
 {
