+++ packages/api/AlvTime.Business/TimeRegistration/TimeEntries/TimeEntryWithCompRateDto.cs
@@ -1,13 +1,13 @@
 ﻿using System;
 
-namespace AlvTime.Business.TimeEntries;
+namespace AlvTime.Business.TimeRegistration.TimeEntries;
 
-public class TimeEntryResponseDto
+public class TimeEntryWithCompRateDto
 {
     public int User { get; set; }
-    public string UserEmail { get; set; }
     public int Id { get; set; }
     public DateTime Date { get; set; }
     public decimal Value { get; set; }
     public int TaskId { get; set; }
+    public decimal CompensationRate { get; set; }
 }
