+++ packages/api/AlvTimeWebApi/Controllers/HolidayController.cs
@@ -1,5 +1,4 @@
 ﻿using AlvTime.Business.Options;
-using AlvTime.Business.TimeEntries;
 using AlvTimeWebApi.Controllers.Utils;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
