+++ packages/api/AlvTimeWebApi/Controllers/Utils/UserContext.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
 using AlvTime.Business.Interfaces;
-using AlvTime.Business.Models;
 using AlvTime.Business.Users;
 using Microsoft.AspNetCore.Http;
