+++ packages/api/Tests/UnitTests/Overtime/OvertimeTests.cs
@@ -6,8 +6,8 @@ using AlvTime.Business.FlexiHours;
 using AlvTime.Business.Interfaces;
 using AlvTime.Business.Options;
 using AlvTime.Business.Payouts;
-using AlvTime.Business.TimeEntries;
 using AlvTime.Business.TimeRegistration;
+using AlvTime.Business.TimeRegistration.TimeEntries;
 using AlvTime.Business.Users;
 using AlvTime.Business.Utils;
 using AlvTime.Persistence.DatabaseModels;
@@ -50,7 +50,7 @@ public class OvertimeTests
 
         _userContextMock = new Mock<IUserContext>();
 
-        var user = new AlvTime.Business.Models.User
+        var user = new AlvTime.Business.Users.User
         {
             Id = 1,
             Email = "someone@alv.no",
@@ -608,7 +608,7 @@ public class OvertimeTests
         await _timeRegistrationService.UpsertTimeEntry(new List<CreateTimeEntryDto>
             {new() {Date = timeEntry1.entry.Date, Value = timeEntry1.entry.Value, TaskId = timeEntry1.taskId}});
             
-        var user2 = new AlvTime.Business.Models.User
+        var user2 = new AlvTime.Business.Users.User
         {
             Id = 2,
             Email = "someone_else@alv.no",
